react-app: split portfolio and stakes query hook refs #172
