+++ react-app/src/components/PortfolioScreen/PortfolioScreen.tsx
@@ -15,7 +15,7 @@ import ProposalHistory, {
   PROPOSAL_HISTORY_PAGE_SIZE,
 } from "../ProposalHistory/ProposalHistory";
 import { useProposalHistory } from "../ProposalHistory/ProposalHistoryAPI";
-import { usePortfolioQuery } from "./PortfolioScreenAPI";
+import { usePortfolioQuery, useStakesQuery } from "./PortfolioScreenAPI";
 import PortfolioPanel from "./PortfolioPanel";
 import StakesPanel from "./StakesPanel";
 
@@ -23,12 +23,15 @@ const PortfolioScreen: React.FC = () => {
   const { address: addressFromUrl } = useParams();
   const navigate = useNavigate();
 
+  const { requestState: portfolioRequestState, fetch: fetchPortfolio } =
+    usePortfolioQuery();
+
   const {
-    requestState: portfolioRequestState,
-    fetch: fetchPortfolio,
-    stakesOrder,
-    setStakesOrder,
-  } = usePortfolioQuery();
+    requestState: stakesRequestState,
+    fetch: fetchStakes,
+    order: stakesOrder,
+    setOrder: setStakesOrder,
+  } = useStakesQuery();
 
   const {
     selectedTab,
@@ -86,8 +89,10 @@ const PortfolioScreen: React.FC = () => {
     if (address) {
       // eslint-disable-next-line @typescript-eslint/no-floating-promises
       fetchPortfolio(address);
+      // eslint-disable-next-line @typescript-eslint/no-floating-promises
+      fetchStakes(address);
     }
-  }, [address, fetchPortfolio]);
+  }, [address, fetchPortfolio, fetchStakes]);
 
   useEffect(() => {
     if (address) {
@@ -105,7 +110,7 @@ const PortfolioScreen: React.FC = () => {
     <div className={cn("flex", "flex-col")}>
       {isRequestStateLoaded(portfolioRequestState) ? (
         <PortfolioPanel
-          portfolio={portfolioRequestState.data.portfolio}
+          portfolio={portfolioRequestState.data}
           isYourPortfolio={isYourPortfolio}
         />
       ) : (
@@ -115,10 +120,10 @@ const PortfolioScreen: React.FC = () => {
       )}
 
       <StakesPanel
-        isLoading={!isRequestStateLoaded(portfolioRequestState)}
+        isLoading={!isRequestStateLoaded(stakesRequestState)}
         stakes={
-          isRequestStateLoaded(portfolioRequestState)
-            ? portfolioRequestState.data.stakes
+          isRequestStateLoaded(stakesRequestState)
+            ? stakesRequestState.data
             : null
         }
         isYourPortfolio={isYourPortfolio}
@@ -127,7 +132,8 @@ const PortfolioScreen: React.FC = () => {
       />
 
       {isRequestStateLoaded(proposalHistoryRequestState) &&
-      isRequestStateLoaded(portfolioRequestState) ? (
+      isRequestStateLoaded(portfolioRequestState) &&
+      isRequestStateLoaded(stakesRequestState) ? (
         <ProposalHistory
           data={proposalHistoryRequestState.data}
           selectedTab={selectedTab}
